requirements.txt
bluepy==1.3.0
btlewrap==0.0.8
mitemp-bt==0.0.3
paho-mqtt==1.5.0
